Update README.md
Privacy policy
Information collection
This Chrome extension does not collect any personal or sensitive user data. No information that could be used to identify you, such as your name, email address, browsing history, or location, is gathered, stored, or transmitted. 
Data usage and sharing
Since no data is collected, none of your data is used for any purpose, including personalization, analytics, or advertising. No data is shared with third parties because no data is collected. 
No storage or transmission
This extension operates locally within your browser and does not store any information on external servers or transmit any data outside your device. Any data used by the extension for its functionality is processed entirely within your browser and is not retained after your browsing session or upon uninstallation of the extension. 
Changes to this privacy policy
This Privacy Policy may be updated to reflect changes in practices or for other operational, legal, or regulatory reasons. Any changes will be posted on this page, and the "Effective Date" at the top of the policy will be updated. You are encouraged to review this Privacy Policy periodically.